requirements.txt
fastapi
uvicorn[standard]
requests
websocket-client
pyyaml
pandas
python-dotenv